fix compilation

+++ src/sdam/description/topology/server_selection/test/in_window.rs
@@ -181,14 +181,14 @@ async fn load_balancing_test() {
         assert!(
             share_of_selections <= max_share,
             "expected no more than {}% of selections, instead got {}%",
-            max_share * 100,
-            share_of_selections * 100
+            (max_share * 100.0) as u32,
+            (share_of_selections * 100.0) as u32
         );
         assert!(
             share_of_selections >= min_share,
             "expected at least {}% of selections, instead got {}%",
-            min_share * 100,
-            share_of_selections * 100
+            (min_share * 100.0) as u32,
+            (share_of_selections * 100.0) as u32
         );
     }
